KOM 20 + 94-95 Progress
Troops spawn to inputs, no script asigned to them so far. So they just fall to the ground and don't move
