--- Assets/DefenderSpawner.cs.meta
@@ -1,2 +1,0 @@
-fileFormatVersion: 2
-guid: 3948833f02941f24cb5f630361d58574
