+++ Assets/TroopFolder.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 986b4679929d032409e93ee798f03f08
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
