+++ Assets/TroopFolder/Troop Prefabs.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 38af221b334255c469336a9ea438193a
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
