+++ Assets/TroopFolder/Troop Prefabs/HeavyAirTroop.prefab.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: 6041eae1d5edcb24693fe16f9d1d7f4a
+PrefabImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
