+++ Assets/TroopFolder/Troop Prefabs/LightAirTroop.prefab.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: b7eccd30b2789d84faf1ecd868ecb787
+PrefabImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
