+++ Assets/TroopFolder/Troop Prefabs/ScoutTroop.prefab.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: 608604b1cb0eece4886a317c05ff1751
+PrefabImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
