+++ Assets/TroopFolder/Troop Prefabs/ScoutTroop2.prefab.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: d5888b11ecbb70742a86edcb78f91dc1
+PrefabImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
