+++ Assets/TroopFolder/Troop Prefabs/TankTroop.prefab.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: 8458fa422e920b54799929408c32832d
+PrefabImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
