+++ Assets/TroopFolder/Troop Scripts.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: bd9dd59a6f3289f4484d6239ae192d8a
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
